# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface PlatformStats {
  students: number;
  units: number;
  assessments: number;
  assignments: number;
  certificates: number;
}

export interface PlatformSettings {
  platformName: string;
  teacherName: string;
  teacherBio: string;
  teacherImageUrl: string;
  signatureUrl: string;
  accentColor: string;
  semester?: string;
  genderSplit?: boolean;
}

export interface PlatformOverview {
  platformName: string;
  teacherName: string;
  tagline: string;
  description: string;
  stats: PlatformStats;
  settings: PlatformSettings;
}

export interface PlatformSettingsUpdate {
  platformName?: string;
  teacherName?: string;
  teacherBio?: string;
  teacherImageUrl?: string;
  signatureUrl?: string;
  accentColor?: string;
  semester?: string;
  genderSplit?: boolean;
}

export interface Student {
  id: string;
  name: string;
  email: string;
  school: string;
  branch: string;
  grade: string;
  section: string;
  gender: string;
  progress: number;
  status: string;
  avatarUrl: string;
}

export interface Course {
  id: string;
  title: string;
  description: string;
  lessons: number;
  duration: string;
  progress: number;
  color: string;
  icon: string;
}

export interface Announcement {
  id: string;
  title: string;
  body: string;
  date: string;
  type: string;
  audience: string;
}

export interface StudentDashboard {
  student: Student;
  progress: number;
  completedLessons: number;
  nextUp: Course[];
  announcements: Announcement[];
}

export interface TeacherStats {
  students: number;
  activeStudents: number;
  schools: number;
  units: number;
  assignments: number;
  assessments: number;
  averageScore: number;
  certificates: number;
}

export interface ActivityPoint {
  label: string;
  value: number;
}

export interface Assignment {
  id: string;
  title: string;
  description: string;
  unit: string;
  courseId?: string | null;
  dueDate: string;
  status: string;
  points: number;
  grade?: string;
  section?: string;
}

export interface TeacherDashboard {
  stats: TeacherStats;
  weeklyActivity: ActivityPoint[];
  recentStudents: Student[];
  pendingAssignments: Assignment[];
}

export interface Assessment {
  id: string;
  title: string;
  questions: number;
  duration: string;
  score: number;
  status: string;
  date: string;
}

export interface AuthRegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  fullName: string;
  studentNumber?: string;
  school?: string;
  branch?: string;
  grade?: string;
  section?: string;
  gender?: string;
  phone?: string;
}

export interface AuthLoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}

export interface AuthEmailInput {
  email: string;
}

export interface AuthUser {
  id: string;
  email: string;
  fullName: string;
}

export interface AuthResponse {
  message: string;
  user: AuthUser;
}

export interface AuthMessage {
  message: string;
}

export interface AuthOAuthResponse {
  url: string;
}