# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: registerAccount
      tags: [auth]
      summary: Create a student account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRegisterInput"
      responses:
        "200":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Invalid account data
  /auth/login:
    post:
      operationId: loginAccount
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthLoginInput"
      responses:
        "200":
          description: Signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
  /auth/forgot-password:
    post:
      operationId: requestPasswordReset
      tags: [auth]
      summary: Request a password reset email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthEmailInput"
      responses:
        "200":
          description: Reset email requested
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthMessage"
...
components:
  schemas:
...
    AuthRegisterInput:
      type: object
      required: [email, password, fullName]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
        fullName: { type: string, minLength: 2 }
        studentNumber: { type: string }
        school: { type: string }
        branch: { type: string }
        grade: { type: string }
        section: { type: string }
        gender: { type: string }
        phone: { type: string }
    AuthLoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 6 }
    AuthEmailInput:
      type: object
      required: [email]
      properties:
        email: { type: string, format: email }
    AuthUser:
      type: object
      required: [id, email, fullName]
      properties:
        id: { type: string }
        email: { type: string }
        fullName: { type: string }
    AuthResponse:
      type: object
      required: [message, user]
      properties:
        message: { type: string }
        user:
          $ref: "#/components/schemas/AuthUser"
    AuthMessage:
      type: object
      required: [message]
      properties:
        message: { type: string }
...