# lib/api-zod/src/generated/types/authLoginInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthLoginInput {
  email: string;
  /** @minLength 6 */
  password: string;
}