# lib/api-zod/src/generated/types/authRegisterInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthRegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 2 */
  fullName: string;
  studentNumber?: string;
  school?: string;
  branch?: string;
  grade?: string;
  section?: string;
  gender?: string;
  phone?: string;
}